feat. TangibleAsset 문제
추상클래스, 상속, 인터페이스, lmplements, get, set
