+++ lib/24_03_13/asset.dart
@@ -1,0 +1,6 @@
+abstract class Asset {   //제일 상단에 있을 추상클래스임
+  String name;
+  int price;
+
+  Asset({required this.name, required this.price});
+}
