+++ lib/24_03_13/thing.dart
@@ -1,0 +1,5 @@
+abstract interface class Thing {
+  //thing이라는 것은 그냥 인터페이스
+  void set weight(double weight); // 무게를 얻을 수 있도록 set값, get값 갖도록
+  double get weight;
+}
